# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ScanRay Sonar API
  description: |
    Single-pane-of-visibility API for the ScanRay Sonar platform.
    Phase 1 surface: authentication, sites, users, agents, appliances,
    plus health and version endpoints.

    All non-auth endpoints require `Authorization: Bearer <jwt>`.
    Mutating endpoints additionally require RBAC role >= the level
    documented per-operation.
  version: "2026.5.5.9"
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Versioned API root

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---- Reusable schemas -------------------------------------------------
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, examples: ["unauthorized"] }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Agent:
      type: object
      required: [id, siteId, hostname, os, isActive, createdAt]
      properties:
        id: { type: string, format: uuid }
        siteId: { type: string, format: uuid }
        hostname: { type: string }
        fingerprint: { type: string, nullable: true }
        os: { type: string }
        osVersion: { type: string }
        agentVersion: { type: string }
        enrolledAt: { type: string, format: date-time, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        isActive: { type: boolean }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /agents:
    get:
      summary: List agents (filter by site)
      parameters:
        - in: query
          name: siteId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Agent" }
...
  /agents/{id}:
    get:
      summary: Get a single agent with its latest snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Agent"
                  - type: object
                    properties:
                      lastMetrics:
                        type: object
                        description: |
                          Verbatim last snapshot the probe sent (Snapshot v2). Includes
                          a `conversations` array of aggregated active TCP/UDP peer
                          pairs (direction, remoteIp, remoteHost, remotePort, localPort,
                          state, pid, processName, count) for the "talking to" view.
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Remove an agent (siteadmin+)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...